linebot_test/requirements.txt
Flask
python-dotenv
gunicorn
google-generativeai
line-bot-sdk
requests
openai